+++ src/plone.server/CHANGELOG.rst
@@ -1,11 +1,14 @@
 1.0a7 (unreleased)
 ------------------
 
+- add jwt token validator
+  [vangheem]
+
 - Add to finalize an AsyncUtil when its finishing the software
   [ramon]
 
 - Remove `AUTH_USER_PLUGINS` and `AUTH_EXTRACTION_PLUGINS`. Authentication now
-  consists of auth policies, user identifiers and token checkers.
+  consists of auth extractors, user identifiers and token validators.
   [vangheem]
 
 - Correctly check parent object for allowed addable types